+++ graspit-reporters/packages/graspit-dashboard/CHANGELOG.md
@@ -1,5 +1,17 @@
 # graspit
 
+## 3.2.0
+
+### Minor Changes
+
+- added support for links and assertions
+
+## 3.1.6
+
+### Patch Changes
+
+- handled permission and attached links
+
 ## 3.1.5
 
 ### Patch Changes